gRPC server returns PredicitionDto

+++ ML-model/dto/predictionDto.py
@@ -3,4 +3,4 @@ from dataclasses import dataclass
 @dataclass
 class PredictionDto:
     activity: str
-    accuracy: int
+    accuracy: float
